backend/requirements.txt
Django==5.2.3
django-cors-headers==4.7.0
djangorestframework==3.16.0
requests==2.32.4
